pyproject.toml
[tool.poetry]
name = "icon-rhizome-dev"
version = "0.1.0"
description = ""
authors = ["Brian Li <brian@brianli.com>"]
readme = "README.md"
packages = [{include = "icon_rhizome_dev"}]

[tool.poetry.dependencies]
python = "^3.11"
iconsdk = "^2.2.0"
starlite = {extras = ["redis"], version = "^1.45.1"}
python-dotenv = "^0.21.0"
jinja2 = "^3.1.2"
jinja2-fragments = "^0.3.0"
uvicorn = "^0.20.0"
httpx = "^0.23.1"
orjson = "^3.8.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
